fix of paramset: accidently used size_t 	modified:   cvt/vision/slam/stereo/PatchStereoInit.cpp 	modified:   cvt/vision/slam/stereo/PatchStereoInit.h

+++ cvt/vision/slam/stereo/PatchStereoInit.h
@@ -32,8 +32,8 @@ namespace cvt {
 
             struct Parameters
             {
-                size_t   maxNewFeatures;
-                size_t   maxSAD;
+                uint32_t maxNewFeatures;
+                uint64_t maxSAD;
                 float    maxEpilineDistance;
                 float    maxReprojectionError;
                 float    maxDepth;
